refactor(api): move /api/admin/users/{id} route and tests
in src/identity-access-management
