+++ api/lib/application/users/index.js
@@ -1,43 +1,11 @@
 import Joi from 'joi';
 
-import { BadRequestError, sendJsonApiError } from '../../../src/shared/application/http-errors.js';
 import { securityPreHandlers } from '../../../src/shared/application/security-pre-handlers.js';
 import { identifiersType } from '../../../src/shared/domain/types/identifiers-type.js';
 import { userController } from './user-controller.js';
 
 const register = async function (server) {
   const adminRoutes = [
-    {
-      method: 'GET',
-      path: '/api/admin/users/{id}',
-      config: {
-        validate: {
-          params: Joi.object({
-            id: identifiersType.userId,
-          }),
-          failAction: (request, h) => {
-            return sendJsonApiError(new BadRequestError("L'identifiant de l'utilisateur n'est pas au bon format."), h);
-          },
-        },
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleCertif,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-                securityPreHandlers.checkAdminMemberHasRoleMetier,
-              ])(request, h),
-          },
-        ],
-        handler: userController.getUserDetailsForAdmin,
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs administrateurs**\n' +
-            "- Elle permet de récupérer le détail d'un utilisateur dans un contexte d'administration",
-        ],
-        tags: ['api', 'admin', 'user'],
-      },
-    },
     {
       method: 'GET',
       path: '/api/admin/users/{id}/organizations',
