+++ api/tests/identity-access-management/integration/application/user/user.admin.route.test.js
@@ -1,3 +1,4 @@
+import { userController } from '../../../../../lib/application/users/user-controller.js';
 import { identityAccessManagementRoutes } from '../../../../../src/identity-access-management/application/routes.js';
 import { userAdminController } from '../../../../../src/identity-access-management/application/user/user.admin.controller.js';
 import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
@@ -136,6 +137,61 @@ describe('Integration | Identity Access Management | Application | Route | Admin
 
       // when
       const response = await httpTestServer.request('PATCH', url);
+
+      // then
+      expect(response.statusCode).to.equal(400);
+    });
+  });
+
+  describe('GET /api/admin/users/{id}', function () {
+    it('returns an HTTP status code 200', async function () {
+      // given
+      sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').returns(() => true);
+      sinon.stub(userController, 'getUserDetailsForAdmin').resolves('ok');
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/admin/users/8');
+
+      // then
+      expect(response.statusCode).to.equal(200);
+      sinon.assert.calledOnce(securityPreHandlers.hasAtLeastOneAccessOf);
+      sinon.assert.calledOnce(userController.getUserDetailsForAdmin);
+    });
+
+    it('returns an HTTP status code 403', async function () {
+      // given
+      sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').returns((request, h) =>
+        h
+          .response({ errors: new Error('') })
+          .code(403)
+          .takeover(),
+      );
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/admin/users/8');
+
+      // then
+      expect(response.statusCode).to.equal(403);
+      sinon.assert.calledOnce(securityPreHandlers.hasAtLeastOneAccessOf);
+    });
+
+    it('returns BAD_REQUEST (400) when id in param is not a number"', async function () {
+      // given
+      const url = '/api/admin/users/NOT_A_NUMBER';
+
+      // when
+      const response = await httpTestServer.request('GET', url);
+
+      // then
+      expect(response.statusCode).to.equal(400);
+    });
+
+    it('returns BAD_REQUEST (400) when id in params is out of range"', async function () {
+      // given
+      const url = '/api/admin/users/0';
+
+      // when
+      const response = await httpTestServer.request('GET', url);
 
       // then
       expect(response.statusCode).to.equal(400);
